everything working

+++ src/app/purchasesPlace/purchases/purchases.component.ts
@@ -70,9 +70,15 @@ showDiv = {
 
 
   deletePurchases(id: number){
+    if(confirm("Are you sure you want to delete this purchase?") == true){
     this.purchaseService.deletePurchases(id).subscribe(data =>{
       this.ngOnInit();
       })
+
+    }
+      else{
+      
+      }
     }
 
     onPageChanged(event: any){
